+++ test/audits/cdn-logs-report/query-builder.test.js
@@ -42,15 +42,13 @@ describe('CDN Logs Query Builder', () => {
         getBaseURL: () => 'https://adobe.com',
         getConfig: () => ({
           getLlmoDataFolder: () => null,
-          getCdnLogsConfig: () => ({
-            filters: [{
-              value: [
-                'www.another.com',
-              ],
-              key: 'host',
-            },
+          getLlmoCdnlogsFilter: () => [{
+            value: [
+              'www.another.com',
             ],
-          }),
+            key: 'host',
+          }],
+          getLlmoCdnBucketConfig: () => ({ orgId: 'test-org-id' }),
         }),
       },
     };
@@ -72,9 +70,20 @@ describe('CDN Logs Query Builder', () => {
   it('handles site filters correctly', async () => {
     mockOptions.site.getConfig = () => ({
       getLlmoDataFolder: () => null,
-      getCdnLogsConfig: () => ({
-        filters: [{ value: ['test'], key: 'url' }],
-      }),
+      getLlmoCdnlogsFilter: () => [{ value: ['test'], key: 'url' }],
+    });
+
+    const query = await weeklyBreakdownQueries.createAgenticReportQuery(mockOptions);
+
+    expect(query).to.include("(REGEXP_LIKE(url, '(?i)(test)'))");
+  });
+
+  it('handles llmo cdn logs site filters correctly', async () => {
+    mockOptions.site.getConfig = () => ({
+      getLlmoDataFolder: () => null,
+      getLlmoCdnlogsFilter: () => (
+        [{ value: ['test'], key: 'url' }]
+      ),
     });
 
     const query = await weeklyBreakdownQueries.createAgenticReportQuery(mockOptions);
@@ -126,7 +135,7 @@ describe('CDN Logs Query Builder', () => {
 
     mockOptions.site.getConfig = () => ({
       getLlmoDataFolder: () => null,
-      getCdnLogsConfig: () => ({}),
+      getLlmoCdnlogsFilter: () => [],
     });
 
     const query = await localQueries.createAgenticReportQuery(mockOptions);